Move map, x-section and widget components to subfolders
