+++ src/components/app.tsx
@@ -2,10 +2,9 @@ import { inject, observer } from "mobx-react";
 import * as React from "react";
 import { LineChart, Line, CartesianGrid, XAxis, YAxis } from "recharts";
 import { BaseComponent, IBaseProps } from "./base";
-import { MapComponent, Scenario } from "./map-component";
+import { MapComponent, Scenario } from "./map/map-component";
 import { LogComponent } from "./log-component";
-import { CrossSectionComponent } from "./cross-section-component";
-import * as Maps from "./../assets/maps/maps.json";
+import { CrossSectionComponent } from "./crosssection/cross-section-component";
 import * as Scenarios from "./../assets/maps/scenarios.json";
 import * as BlocklyAuthoring from "./../assets/blockly-authoring/index.json";
 
@@ -19,7 +18,7 @@ import SyntaxHighlighter from "react-syntax-highlighter";
 import Controls from "./controls";
 import RunButtons from "./run-buttons";
 import { Footer, TabContent } from "./styled-containers";
-import WidgetPanel from "./widget-panel";
+import WidgetPanel from "./widgets/widget-panel";
 import screenfull from "screenfull";
 import ResizeObserver from "react-resize-observer";
 import AuthoringMenu from "./authoring-menu";
