+++ src/components/crosssection/cross-section-component.tsx
@@ -1,11 +1,11 @@
 import * as React from "react";
-import { ICanvasShape } from "../interfaces";
+import { ICanvasShape } from "../../interfaces";
 import styled from "styled-components";
 import { Stage } from "@inlet/react-pixi";
 import { PixiTephraCrossSection } from "./pixi-tephra-cross-section";
 import * as Color from "color";
 import { inject, observer } from "mobx-react";
-import { BaseComponent, IBaseProps } from "./base";
+import { BaseComponent, IBaseProps } from "../base";
 const CanvDiv = styled.div`
   border: 0px solid black; border-radius: 0px;
 `;
