+++ src/components/map/map-component.tsx
@@ -3,20 +3,20 @@ import * as L from "leaflet";
 
 import { Map as LeafletMap, TileLayer, Marker, Popup, ScaleControl, Pane } from "react-leaflet";
 import "leaflet/dist/leaflet.css";
-import "../css/map-component.css";
-import { iconVolcano, iconMarker, getCachedDivIcon } from "./icons";
+import "../../css/map-component.css";
+import { iconVolcano, iconMarker, getCachedDivIcon } from "../icons";
 
-import { CityType  } from "../stores/simulation-store";
-import * as Scenarios from "./../assets/maps/scenarios.json";
+import { CityType  } from "../../stores/simulation-store";
+import * as Scenarios from "../../assets/maps/scenarios.json";
 import styled from "styled-components";
 import { observer, inject } from "mobx-react";
-import { BaseComponent, IBaseProps } from "./base";
-import { CrossSectionDrawLayer } from "./cross-section-draw-layer";
-import { LocalToLatLng } from "../utilities/coordinateSpaceConversion";
+import { BaseComponent, IBaseProps } from "../base";
+import { CrossSectionDrawLayer } from "./layers/cross-section-draw-layer";
+import { LocalToLatLng } from "../../utilities/coordinateSpaceConversion";
 import { MapTephraThicknessLayer } from "./map-tephra-thickness-layer";
-import { OverlayControls } from "./overlay-controls";
-import { RulerDrawLayer } from "./ruler-draw-layer";
-import { RightSectionTypes } from "./tabs";
+import { OverlayControls } from "../overlay-controls";
+import { RulerDrawLayer } from "./layers/ruler-draw-layer";
+import { RightSectionTypes } from "../tabs";
 import KeyButton from "./map-key-button";
 import CompassComponent from "./map-compass";
 import TephraLegendComponent from "./map-tephra-legend";
