+++ src/components/widgets/ejected-volume-widget.tsx
@@ -1,12 +1,12 @@
 import * as React from "react";
 import { PureComponent } from "react";
 import styled from "styled-components";
-import { Icon } from "./icon";
-import BoxBackIcon from "../assets/widget-icons/ejected-volume-box-back.svg";
-import BoxFrontIcon from "../assets/widget-icons/ejected-volume-box-front.svg";
-import BoxTopIcon from "../assets/widget-icons/ejected-volume-box-top.svg";
-import { ValueContainer, ValueOutput, IconContainer } from "./styled-containers";
-import { WidgetPanelTypes, kWidgetPanelInfo } from "../utilities/widget";
+import { Icon } from "../icon";
+import BoxBackIcon from "../../assets/widget-icons/ejected-volume-box-back.svg";
+import BoxFrontIcon from "../../assets/widget-icons/ejected-volume-box-front.svg";
+import BoxTopIcon from "../../assets/widget-icons/ejected-volume-box-top.svg";
+import { ValueContainer, ValueOutput, IconContainer } from "../styled-containers";
+import { WidgetPanelTypes, kWidgetPanelInfo } from "../../utilities/widget";
 
 interface BoxProps {
   height?: number;
