+++ src/components/widgets/vei-widget.tsx
@@ -1,27 +1,27 @@
 import * as React from "react";
 import { PureComponent } from "react";
 import styled from "styled-components";
-import { Icon } from "./icon";
-import VEI1Orange from "../assets/widget-icons/vei-1-orange.svg";
-import VEI2Orange from "../assets/widget-icons/vei-2-orange.svg";
-import VEI3Orange from "../assets/widget-icons/vei-3-orange.svg";
-import VEI4Orange from "../assets/widget-icons/vei-4-orange.svg";
-import VEI5Orange from "../assets/widget-icons/vei-5-orange.svg";
-import VEI6Orange from "../assets/widget-icons/vei-6-orange.svg";
-import VEI7Orange from "../assets/widget-icons/vei-7-orange.svg";
-import VEI8Orange from "../assets/widget-icons/vei-8-orange.svg";
-import VEI1Blue from "../assets/widget-icons/vei-1-blue.svg";
-import VEI2Blue from "../assets/widget-icons/vei-2-blue.svg";
-import VEI3Blue from "../assets/widget-icons/vei-3-blue.svg";
-import VEI4Blue from "../assets/widget-icons/vei-4-blue.svg";
-import VEI5Blue from "../assets/widget-icons/vei-5-blue.svg";
-import VEI6Blue from "../assets/widget-icons/vei-6-blue.svg";
-import VEI7Blue from "../assets/widget-icons/vei-7-blue.svg";
-import VEI8Blue from "../assets/widget-icons/vei-8-blue.svg";
+import { Icon } from "../icon";
+import VEI1Orange from "../../assets/widget-icons/vei-1-orange.svg";
+import VEI2Orange from "../../assets/widget-icons/vei-2-orange.svg";
+import VEI3Orange from "../../assets/widget-icons/vei-3-orange.svg";
+import VEI4Orange from "../../assets/widget-icons/vei-4-orange.svg";
+import VEI5Orange from "../../assets/widget-icons/vei-5-orange.svg";
+import VEI6Orange from "../../assets/widget-icons/vei-6-orange.svg";
+import VEI7Orange from "../../assets/widget-icons/vei-7-orange.svg";
+import VEI8Orange from "../../assets/widget-icons/vei-8-orange.svg";
+import VEI1Blue from "../../assets/widget-icons/vei-1-blue.svg";
+import VEI2Blue from "../../assets/widget-icons/vei-2-blue.svg";
+import VEI3Blue from "../../assets/widget-icons/vei-3-blue.svg";
+import VEI4Blue from "../../assets/widget-icons/vei-4-blue.svg";
+import VEI5Blue from "../../assets/widget-icons/vei-5-blue.svg";
+import VEI6Blue from "../../assets/widget-icons/vei-6-blue.svg";
+import VEI7Blue from "../../assets/widget-icons/vei-7-blue.svg";
+import VEI8Blue from "../../assets/widget-icons/vei-8-blue.svg";
 import { HorizontalContainer, VerticalContainer,
-         ValueContainer, ValueOutput, IconContainer } from "./styled-containers";
-import { kVEIIndexInfo } from "../utilities/vei";
-import { WidgetPanelTypes, kWidgetPanelInfo } from "../utilities/widget";
+         ValueContainer, ValueOutput, IconContainer } from "../styled-containers";
+import { kVEIIndexInfo } from "../../utilities/vei";
+import { WidgetPanelTypes, kWidgetPanelInfo } from "../../utilities/widget";
 
 interface VEIlabelProps {
   color?: string;
